+++ nix/derivation.nix
@@ -1,7 +1,6 @@
 {
   lib,
   stdenv,
-  curl,
   meson,
   ninja,
   pciutils,